as libten uses fcntl, its third parameter is int, not long

+++ include/ten/descriptors.hh
@@ -77,8 +77,8 @@ struct fd_base {
         static_assert(!std::is_same<T, bool>::value, "fd == bool");
     }
 
-    int fcntl(int cmd) { return ::fcntl(fd, cmd); }
-    int fcntl(int cmd, long arg) { return ::fcntl(fd, cmd, arg); }
+    int fcntl(int cmd)              { return ::fcntl(fd, cmd); }
+    int fcntl(int cmd, int arg)     { return ::fcntl(fd, cmd, arg); }
     int fcntl(int cmd, flock *lock) { return ::fcntl(fd, cmd, lock); }
 
     //! make fd nonblocking by setting O_NONBLOCK flag
